Crate ftp_lib

ftp_lib

ftp_lib is a FTP implementation for a school project

Example:

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

// Create a directory
ftp_conn.mkdir("test_dir").unwrap();

// Move into the newly created directory and print the working directory
ftp_conn.cd("test_dir").unwrap();
println!("PWD: {:?}", ftp_conn.pwd().unwrap());

// Return to / from /test_dir and print the working directory
ftp_conn.cdup().unwrap();
println!("PWD: {:?}", ftp_conn.pwd().unwrap());

// Remove the directory created
ftp_conn.rmdir("test_dir").unwrap();

ftp_conn.quit().unwrap();

Modules

error
Module containing all errors returned by ftp_lib.

filestructure
Module for handling the outputs of ftp_conn.list()

status
Module containing all the status codes required to handle responses.

Structs

FtpConnection
The main type used for communication with the FTP server.

FtpResponse
Data structure that contains a response from the FTP server.

Enums

FtpState
The state of the ftp connection, connected or not