Module ftp_lib::status::ftp_status

Module responsible for containing the many constants needed for holding FTP response codes.

Constants

ACCOUNT_REQUIRED_LOGIN
Need account for login.

ACCOUNT_REQUIRED_STORING
Need account for storing files.

ACTION_ABORTED_PROCESSING
Requested action aborted. Local error in processing.

ACTION_ABORTED_UNKOWN_PAGE
Requested action aborted. Page type unknown.

ACTION_NOT_TAKEN
Requested action not taken. File unavailable (e.g., file not found, no access).

BAD_COMMAND_SEQUENCE
Bad sequence of commands.

COMMAND_NOT_IMPLEMENTED
Command not implemented.

COMMAND_NOT_IMPLEMENTED_PARAMETER
Command not implemented for that parameter.

COMMAND_NOT_IMPLEMENTED_UNNECESARY
Command not implemented, superfluous at this site.

COMMAND_OKAY
Command okay.

DATA_CANNOT_CONNECT
Can't open data connection.

DATA_CLOSED_ABORTING
Connection closed; transfer aborted.

DATA_CLOSING
Closing data connection. Requested file action successful (for example, file transfer or file abort).

DATA_NOT_TRANSFERING
Data connection open; no transfer in progress.

DATA_TRANSFER_STARTING
Data connection already open; transfer starting.

DIRECTORY_CREATED
"PATHNAME" created.

DIRECTORY_STATUS
Directory status.

ENTERING_PASSIVE
Entering Passive Mode (h1,h2,h3,h4,p1,p2).

FILE_ACTION_COMPLETE
Requested file action okay, completed.

FILE_ACTION_NOT_TAKEN
File action not taken. File unavailable (e.g., file busy)

FILE_NAME_INVALID
Requsted action not taken. File name not allowed.

FILE_NEED_INFORMATION
Requested file action pending further information.

FILE_OPENING_DATA
File status okay; about to open the data connection.

FILE_STATUS
File status.

HELP_MESSAGE
Help message. On how to use the server or the meaning of a particular non-standard command. This reply is useful only to the human user.

INSUFFICIENT_ALLOCATED_STORAGE
Requested file action aborted. Exceeded storage allocation (for current directory or dataset).

INSUFFICIENT_STORAGE
Requested action not taken. Insufficient storage space in system.

LOGGED_IN
User logged in, proceed.

NOT_LOGGED_IN
Not logged in.

PASSWORD_NEEDED
User name okay, need password.

READY_IN
Service ready in nnn minutes.

RESTART_MARKER_REPLY
Restart marker reply. In this case, the text is exact and not left to the particular implementation; it must read: MARK yyyy = mmmm Wthere yyyy is User-process data stream market, and mmmm server's equivelent market (note the space between markers and "=")

SERVER_CLOSING_CONTROL
Service closing control connection. Logged out if approperiate.

SERVICE_NOT_AVAILABLE
Service not available, closing control connection. This may be a reply to any command if the service knows it must shut down.

SERVICE_READY
Servuce ready for new user.

SYNTAX_ERROR
Syntax error, command unrecognized.

SYNTAX_ERROR_ARGUMENTS
Syntax error in parameters or arguments.

SYSTEM_STATUS
System status, or system help reply.

SYSTEM_TYPE
NAME system type. Where NAME is an official system name from the list in the Assigned Numbers document.