Struct ftp_lib::FtpConnection

pub struct FtpConnection { /* fields omitted */ }

The main type used for communication with the FTP server.

Once calling the constructor a connection to the specified socket address is opened, allowing for login and additional operations to be performed.

Example:

Connection

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap(); // Initiate the connection

ftp_conn.quit().unwrap();

Methods

impl FtpConnection

pub fn connect(connection_addr: SocketAddrV4) -> Result<FtpConnection, FtpError>

Initiates the connection to the FTP server.

After initiating the stream of data to the server, the function waits for a response telling it if the service is available or if the service is not yet ready. If and once the server is ready, the function will return a functional instance of FtpConnection ready for varius operations to be performed.

Example:

Connect to 127.0.0.1 then close the connection.

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap(); // Initiate the connection

ftp_conn.quit().unwrap();

pub fn login(
&mut self,
username: &str,
password: Option<&str>
) -> Result<(), FtpError>

Authenticates the connection to the FTP server.

When connecting to an FTP server the first command expected is the USER command specifying a username. From there a password should then be given logging the user in if their credentials were correct. A failure to authenticate will return an error of the type FtpError::NotLoggedIn.

Example:

Login to 127.0.0.1 then close the connection.

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

// login(username, password)
ftp_conn.login("anonymous", Some("fake@email.service")).unwrap(); // Login to the server

ftp_conn.quit().unwrap();

pub fn quit(&mut self) -> Result<(), FtpError>

Terminates the connection to the FTP server.

As long as no connection errors occur while sending the quit command or shutting down both read and write interfaces, the connection will be severed, despite what the servers response is.

Example:

Connect to localhost then close the connection.

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

// Preform actions with the FTP server

ftp_conn.quit().unwrap(); // End the connection to the server.

pub fn cd(&mut self, path: &str) -> Result<(), FtpError>

Changes the current working directory in the FTP server.

Works similarly to the cd command in shells. On successful completion of this command, pwd should often follow to be absolutely sure the program knows where it currently is, this is due to the fact that the client cannot be certain of how the server will behave in response.

Example:

Connect to 127.0.0.1 then change to the /test_dir directory.

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

ftp_conn.cd("test_dir").unwrap(); // Change the working directory to rfc

ftp_conn.quit().unwrap();

pub fn cdup(&mut self) -> Result<(), FtpError>

Changes the current working directory in the FTP server to the parent directory.

Simmilar to the cd function, but simply goes up to the parent directory. This function is the same as cd .. in most shells.

Example:

Connect to localhost then go into /test_dir and return to /.

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

ftp_conn.cd("test_dir").unwrap();
println!("PWD: {:?}", ftp_conn.pwd());

ftp_conn.cdup().unwrap(); // Return to / from /test_dir
println!("PWD: {:?}", ftp_conn.pwd());

ftp_conn.quit().unwrap();

pub fn pwd(&mut self) -> Result<String, FtpError>

Gets the current working directory on the FTP server.

The working directory is important to keep track of in order To orient the client in the server's file system. The working directory is the default place most commands are executed from unless instructed otherwise.

Example:

Connect to localhost then fetch the current working directory

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

println!("PWD: {:?}", ftp_conn.pwd()); // Fetch the current working directory

ftp_conn.quit().unwrap();

pub fn list(&mut self) -> Result<Vec<DirectoryItem>, FtpError>

Lists files in the current directory

Listing the files in the directory can be a difficult task, if the server being used does not follow the common output, the program may not be able to understand the current directory. If there is an issue, another directory list structure may need to be implemented.

Example:

Lists all files in /

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

println!("{:?}", ftp_conn.list().unwrap()); // Print files in current directory

ftp_conn.quit().unwrap();

pub fn name_list(&mut self) -> Result<Vec<String>, FtpError>

Lists files in the current directory (Names only)

This function has a definite output unlike LIST. However, less detailed information is returned. Only the names of the files are returned from this function.

Example:

Lists all files in /

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

println!("{:?}", ftp_conn.name_list().unwrap()); // Print files in current directory

ftp_conn.quit().unwrap();

pub fn mkdir(&mut self, dir_name: &str) -> Result<(), FtpError>

Creates a new directory on the FTP server.

E## xample: Connect to localhost then create the directory cool_directory

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

ftp_conn.mkdir("cool_directory").unwrap(); // Create the directory cool_directory
println!("{:?}", ftp_conn.list().unwrap()); // Show the files in the current directory

ftp_conn.quit().unwrap();

pub fn rmdir(&mut self, directory: &str) -> Result<(), FtpError>

Removees a directory on the FTP server.

Example:

Connect to localhost then close the connection.

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

println!("{:?}", ftp_conn.list().unwrap()); // Show before
ftp_conn.rmdir("test_dir").unwrap(); // Remove the directory
println!("{:?}", ftp_conn.list().unwrap()); // Show after

ftp_conn.quit().unwrap();

pub fn fetch_file(&mut self, file_name: &str) -> Result<Vec<u8>, FtpError>

Fetches the contents of the specified file

Example:

Connect to localhost then read README.txt

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

println!("{}",
    String::from_utf8_lossy(&ftp_conn.fetch_file("README.txt").unwrap()) // Fetch the contents of the file
);

ftp_conn.quit().unwrap();

pub fn write_file(
&mut self,
file_name: &str,
data: Vec<u8>
) -> Result<(), FtpError>

Removees a file on the FTP server.

Example:

Writes test.txt

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

ftp_conn.write_file("test.txt", "Cool Data here".as_bytes().to_vec()).unwrap();
println!("{}",
    String::from_utf8_lossy(&ftp_conn.fetch_file("test.txt").unwrap())
);

ftp_conn.quit().unwrap();

pub fn rm(&mut self, file: &str) -> Result<(), FtpError>

Removes a file on the FTP server.

Example:

Removes test.txt

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

println!("{:?}", ftp_conn.list().unwrap()); // Show before
ftp_conn.rm("test.txt").unwrap(); // Remove the directory
println!("{:?}", ftp_conn.list().unwrap()); // Show after

ftp_conn.quit().unwrap();

pub fn rename(&mut self, src: &str, dst: &str) -> Result<(), FtpError>

Renames the specified file on the FTP server

Example:

Connect to localhost rename a new file

use ftp_lib::FtpConnection;
use std::net::SocketAddrV4;

let mut ftp_conn = FtpConnection::connect(
    "127.0.0.1:21".parse().unwrap()
).unwrap();

ftp_conn.login("anonymous", Some("fake@email.service")).unwrap();

println!("{:?}", ftp_conn.list().unwrap()); // Display begining file structure
ftp_conn.rename("test.txt", "cool.txt"); // Rename from test.txt to cool.txt
println!("{:?}", ftp_conn.list().unwrap()); // Verify the change

ftp_conn.quit().unwrap();